src/config_loader.py
# config_loader.py
import yaml
from pathlib import Path


def load_config():
    config_path = Path(__file__).parent / "config.yaml"
    with open(config_path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)


# Tworzymy globalną instancję konfiguracji
cfg = load_config()
